AER-4808 - Added a "format" attribute to the declarations of cf_fault_event(), cf_fault_event2(), and cf_fault_event_nostack(), so that the compiler will warn if the format string in a logging call does not match the passed arguments. Also, fixed mismatches between format strings and argument types in logging calls.

+++ as/src/base/batch.c
@@ -658,7 +658,7 @@ as_batch_queue_task(as_transaction* btr)
 
 	if (bproto->sz > PROTO_SIZE_MAX) {
 		cf_warning(AS_BATCH, "can't process message: invalid size %"PRIu64" should be %d or less",
-				bproto->sz, PROTO_SIZE_MAX);
+				(uint64_t)bproto->sz, PROTO_SIZE_MAX);
 		return as_batch_send_error(btr, AS_PROTO_RESULT_FAIL_PARAMETER);
 	}
